+++ test/custom_model_path/models/tableA/generated.rs
@@ -8,13 +8,13 @@ use serde::{Deserialize, Serialize};
 
 type ConnectionType = diesel::r2d2::PooledConnection<diesel::r2d2::ConnectionManager<diesel::PgConnection>>;
 
-#[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable, Selectable)]
+#[derive(Debug, Clone, Serialize, Deserialize, Queryable, Selectable)]
 #[diesel(table_name=tableA, primary_key(_id))]
 pub struct TableA {
     pub _id: i32,
 }
 
-#[derive(Debug, Serialize, Deserialize, Clone, Queryable, Insertable)]
+#[derive(Debug, Clone, Serialize, Deserialize, Insertable)]
 #[diesel(table_name=tableA)]
 pub struct CreateTableA {
     pub _id: i32,